[interoperator] Dynamic Watch
- Do not add dummy instance and binding to apiserver
- Move deleteObject to test
